# spec-api-swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OpenAPI Spec Management API",
    "description": "REST API for managing OpenAPI specifications in the database",
    "version": "1.0.0",
    "contact": {
      "name": "OpenAPI MCP",
      "url": "https://github.com/ubermorgenland/openapi-mcp"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Integrated MCP and Spec Management server"
    }
  ],
  "paths": {
...
    "/specs/{id}": {
      "get": {
        "summary": "Get OpenAPI spec by ID",
        "description": "Retrieve a specific OpenAPI specification by its ID",
        "operationId": "getSpec",
        "parameters": [
          {
            "$ref": "#/components/parameters/SpecId"
          }
        ],
        "responses": {
          "200": {
            "description": "Spec retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "501": {
            "$ref": "#/components/responses/NotImplemented"
          }
        }
      },
      "put": {
        "summary": "Update OpenAPI spec",
        "description": "Update an existing OpenAPI specification",
        "operationId": "updateSpec",
        "parameters": [
          {
            "$ref": "#/components/parameters/SpecId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpecRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spec updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "501": {
            "$ref": "#/components/responses/NotImplemented"
          }
        }
      },
      "delete": {
        "summary": "Delete OpenAPI spec",
        "description": "Permanently delete an OpenAPI specification",
        "operationId": "deleteSpec",
        "parameters": [
          {
            "$ref": "#/components/parameters/SpecId"
          }
        ],
        "responses": {
          "200": {
            "description": "Spec deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "message": "Spec deleted successfully",
                  "data": {
                    "id": 2
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "SpecId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "The ID of the OpenAPI specification",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "example": 1
        }
      }
    },
    "schemas": {
...
      "UpdateSpecRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Human-readable name for the spec"
          },
          "endpoint_path": {
            "type": "string",
            "description": "Endpoint path for MCP server"
          },
          "spec_content": {
            "type": "string",
            "description": "Raw OpenAPI specification content"
          },
          "file_format": {
            "type": "string",
            "enum": ["json", "yaml"],
            "description": "Format of the specification"
          },
          "api_key_token": {
            "type": "string",
            "description": "API key token for this specification"
          },
          "active": {
            "type": "boolean",
            "description": "Whether the spec should be active"
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "data": {
            "type": "object",
            "description": "Response data (varies by endpoint)"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Name is required"
          },
          "code": {
            "type": "integer",
            "example": 400
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Bad Request",
              "message": "Name is required",
              "code": 400
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Not Found",
              "message": "Spec with ID 999 not found",
              "code": 404
            }
          }
        }
      },
...
      "NotImplemented": {
        "description": "Feature not implemented",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Not Implemented",
              "message": "Get spec by ID not implemented yet",
              "code": 501
            }
          }
        }
      }
    }
  }
}